+++ src/main/java/edu/unbosque/Workshop5/jpa/repositories/LibraryRepositoryImpl.java
@@ -1,6 +1,7 @@
 package edu.unbosque.Workshop5.jpa.repositories;
 
 import edu.unbosque.Workshop5.jpa.entities.Author;
+import edu.unbosque.Workshop5.jpa.entities.Edition;
 import edu.unbosque.Workshop5.jpa.entities.Library;
 
 import javax.persistence.EntityManager;
@@ -27,6 +28,12 @@ public class LibraryRepositoryImpl implements LibraryRepository {
     }
 
     @Override
+    public Library findById2(Integer id) {
+        Library library = entityManager.find(Library.class, id);
+        return library;
+    }
+
+    @Override
     public Optional<Library> save(Library library) {
         try {
             entityManager.getTransaction().begin();
@@ -46,7 +53,25 @@ public class LibraryRepositoryImpl implements LibraryRepository {
             try {
 
                 entityManager.getTransaction().begin();
+                entityManager.remove(library);
+                entityManager.getTransaction().commit();
 
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    @Override
+    public void deleteEditionById(Integer id) {
+        Library library = entityManager.find(Library.class, id);
+        if (library != null) {
+            try {
+
+                entityManager.getTransaction().begin();
+                library.getEditions().forEach(edition -> {
+                    edition.getLibraries().remove(library);
+                });
                 entityManager.remove(library);
                 entityManager.getTransaction().commit();
